refactor: isolate legacy mission loading
