+++ packages/daemon/src/tui/mirror/missions-workspace.ts
@@ -10,16 +10,9 @@ import type {
   TaskCardView,
 } from "@tmux-ide/contracts";
 
-import {
-  MissionProjectionError,
-  projectMissionBoard,
-  projectMissionDetail,
-  projectMissionHistory,
-} from "../../lib/mission-projections.ts";
-import { MissionRepository, MissionRepositoryError } from "../../lib/mission-repository.ts";
-import type { ProjectRuntimeRepository } from "../../lib/project-runtime-repository.ts";
 import type { HostedPanelView } from "./panel-host.ts";
 import { findFirstHostedViewForPanel, terminalDisplayWidth } from "./panel-host.ts";
+import { clipTerminal } from "./terminal-text.ts";
 import type { WorkspaceMissionsNavigationState, WorkspaceUiStateV2 } from "./workspace-ui-state.ts";
 import {
   missionsSelection,
@@ -27,6 +20,10 @@ import {
   setMissionsSelection,
   viewStateFor,
 } from "./workspace-ui-state.ts";
+
+// Compatibility for callers that already import mission-specific helpers from
+// this module. Generic surfaces should import from terminal-text directly.
+export { clipTerminal } from "./terminal-text.ts";
 
 export const MISSION_BOARD_COLUMNS = ["planned", "running", "blocked", "review", "done"] as const;
 export const MISSION_COLUMN_LABELS: Readonly<Record<MissionBoardColumn, string>> = {
@@ -280,81 +277,8 @@ export type MissionWorkspaceHit =
   | { kind: "deep-link"; link: MissionDeepLinkKind }
   | null;
 
-export class MissionWorkspaceLoader {
-  #generation = 0;
-  #activeProjectKey: string | null = null;
-
-  begin(
-    projectKey: string | null,
-    priorSnapshot?: MissionWorkspaceSnapshot | null,
-  ): MissionWorkspaceLoadState {
-    this.#generation += 1;
-    this.#activeProjectKey = projectKey;
-    if (projectKey && priorSnapshot?.project.identityKey === projectKey) {
-      return {
-        status: "refreshing",
-        generation: this.#generation,
-        projectKey,
-        snapshot: priorSnapshot,
-      };
-    }
-    return { status: "loading", generation: this.#generation, projectKey };
-  }
-
-  cancel(): void {
-    this.#generation += 1;
-    this.#activeProjectKey = null;
-  }
-
-  accept(
-    generation: number,
-    projectKey: string,
-    snapshot: MissionWorkspaceSnapshot,
-  ): MissionWorkspaceLoadState | null {
-    if (generation !== this.#generation || projectKey !== this.#activeProjectKey) return null;
-    const status =
-      snapshot.board.counts.total === 0 && snapshot.history.length === 0 ? "empty" : "ready";
-    return { status, generation, snapshot };
-  }
-
-  reject(
-    generation: number,
-    projectKey: string | null,
-    error: unknown,
-  ): MissionWorkspaceLoadState | null {
-    if (generation !== this.#generation || projectKey !== this.#activeProjectKey) return null;
-    return { status: "error", generation, projectKey, message: missionErrorMessage(error) };
-  }
-
-  isCurrent(generation: number, projectKey: string | null): boolean {
-    return generation === this.#generation && projectKey === this.#activeProjectKey;
-  }
-}
-
 export function invalidatedMissionWorkspaceLoadState(): MissionWorkspaceLoadState {
   return { status: "loading", generation: 0, projectKey: null };
-}
-
-export function readMissionWorkspace(
-  repository: ProjectRuntimeRepository,
-  selectedMissionId: string | null = null,
-  now: () => Date = () => new Date(),
-): MissionWorkspaceSnapshot {
-  const missions = new MissionRepository(repository);
-  const { history, state } = missions.snapshot();
-  const board = projectMissionBoard(state, history);
-  const completed = projectMissionHistory(state, history);
-  const detail = selectedMissionId ? projectDetailOrNull(state, history, selectedMissionId) : null;
-  return {
-    board,
-    history: completed.map((entry) => detached(entry)),
-    detail: detail ? detached(detail) : null,
-    project: {
-      identityKey: repository.metadata.identityKey,
-      projectRoot: repository.metadata.projectRoot,
-    },
-    loadedAt: now().toISOString(),
-  };
 }
 
 export function defaultMissionWorkspaceModel(
@@ -1288,49 +1212,6 @@ function missionModelToNavigation(model: MissionWorkspaceModel): WorkspaceMissio
   };
 }
 
-export function clipTerminal(text: string, width: number): string {
-  if (width <= 0) return "";
-  if (terminalDisplayWidth(text) <= width) return text;
-  const ellipsis = "…";
-  const limit = Math.max(0, width - terminalDisplayWidth(ellipsis));
-  let out = "";
-  let used = 0;
-  for (const segment of graphemes(text)) {
-    const segmentWidth = terminalDisplayWidth(segment);
-    if (used + segmentWidth > limit) break;
-    out += segment;
-    used += segmentWidth;
-  }
-  return out + ellipsis;
-}
-
-function missionErrorMessage(error: unknown): string {
-  if (error instanceof MissionRepositoryError || error instanceof MissionProjectionError) {
-    return error.message;
-  }
-  if (error instanceof Error) return error.message;
-  return "mission data could not be loaded";
-}
-
-function projectDetailOrNull(
-  state: Parameters<typeof projectMissionDetail>[0],
-  history: Parameters<typeof projectMissionDetail>[1],
-  missionId: MissionCardView["id"],
-): MissionDetailView | null {
-  try {
-    return projectMissionDetail(state, history, missionId);
-  } catch (error) {
-    if (error instanceof MissionProjectionError && error.projectionCode === "MISSION_NOT_FOUND") {
-      return null;
-    }
-    throw error;
-  }
-}
-
-function detached<T>(value: T): T {
-  return JSON.parse(JSON.stringify(value)) as T;
-}
-
 function emptyScrolls(): Record<MissionBoardColumn, number> {
   return { planned: 0, running: 0, blocked: 0, review: 0, done: 0 };
 }
@@ -2112,13 +1993,4 @@ function formatDuration(durationMs: number | null): string {
   if (minutes < 60) return `${minutes}m`;
   const hours = Math.floor(minutes / 60);
   return `${hours}h ${minutes % 60}m`;
-}
-
-function graphemes(text: string): string[] {
-  const Segmenter = Intl.Segmenter;
-  if (Segmenter)
-    return [...new Segmenter(undefined, { granularity: "grapheme" }).segment(text)].map(
-      (entry) => entry.segment,
-    );
-  return [...text];
 }
